requirements.txt
Flask>=2.3.2
flask-cors>=3.0.10
openai>=1.0.0
python-dotenv>=1.0.0